requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.6.2
Django==4.2.13
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
packaging==24.1
pip-chill==1.0.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1